domain/src/test/kotlin/mega/privacy/android/domain/usecase/camerauploads/ListenToNewMediaUseCaseTest.kt
package mega.privacy.android.domain.usecase.camerauploads

import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.runTest
import mega.privacy.android.domain.repository.CameraUploadsRepository
import org.junit.jupiter.api.BeforeAll
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.TestInstance
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.ValueSource
import org.mockito.kotlin.mock
import org.mockito.kotlin.reset
import org.mockito.kotlin.verify

/**
 * Test class for [ListenToNewMediaUseCase]
 */
@ExperimentalCoroutinesApi
@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class ListenToNewMediaUseCaseTest {

    private lateinit var underTest: ListenToNewMediaUseCase

    private val cameraUploadsRepository = mock<CameraUploadsRepository>()

    @BeforeAll
    fun setUp() {
        underTest = ListenToNewMediaUseCase(
            cameraUploadsRepository = cameraUploadsRepository,
        )
    }

    @BeforeEach
    fun resetMocks() {
        reset(cameraUploadsRepository)
    }

    @ParameterizedTest(name = "when forceEnqueue is {0}")
    @ValueSource(booleans = [true, false])
    fun `test that listen to new media is invoked`(forceEnqueue: Boolean) = runTest {
        underTest(forceEnqueue = forceEnqueue)

        verify(cameraUploadsRepository).listenToNewMedia(forceEnqueue = forceEnqueue)
    }
}
